coinstreak , collatz and comma
These three projects are one that drill the fundamentals of Python.

the comma requires you to print string elements of a list as a string with comma and the last item with and <item>.

Coinstreak checks the number of times a streak of 6 heads or tails happen and calculate the probability of occurence.

Collatz is a popular mathematical problem. You can check out Collatz sequence to learn more.